@@ -1,6 +1,6 @@
 {
   "name": "prague",
-  "version": "0.22.4",
+  "version": "0.23.0",
   "description": "EDSL for rules",
   "main": "lib/src/prague.js",
   "types": "lib/src/prague.d.ts",
@@ -15,10 +15,11 @@
     "@types/chai": "^4.1.4",
     "@types/chai-subset": "^1.3.1",
     "@types/mocha": "^5.2.5",
+    "@types/node": "^10.9.2",
     "chai": "^4.1.2",
     "chai-subset": "^1.6.0",
     "mocha": "^5.2.0",
-    "typescript": "^3.0.1"
+    "typescript": "^3.1.0-dev.20180825"
   },
   "dependencies": {
     "rxjs": "6.2.2"
