@@ -1,7 +1,8 @@
 export * from './core';
 export * from './first';
 export * from './pipe';
-export * from './scoring';
+export * from './multiple';
+export * from './scored';
 export * from './matching';
 export * from './actionReference';
 export * from './util';
